+++ docs/Changelog.md
@@ -3,4 +3,4 @@
 ## 0.6.3 (unreleased)
 * hidden [`--integrate-to-system`](Overview.md#bash-completion) argument
 * interfaces migrated to [`mininterface.interfaces`](Interfaces.md) to save around 50 ms starting time due to lazy loading
-* [SubcommandPlaceholder][mininterface.common.SubcommandPlaceholder]
+* [SubcommandPlaceholder][mininterface.subcommands.Command]